- shutup many compiler warnings. #1513
git-svn-id: svn://svn.lighttpd.net/lighttpd/trunk@2048 152afb58-edef-0310-8abb-c4023f1b3aa9

+++ src/log.c
@@ -321,7 +321,7 @@ int log_trace(const char *fmt, ...) {
 		l = vsnprintf(b->ptr, b->size, fmt, ap);
 		va_end(ap);
 
-		if (l > -1 && l < b->size) {
+		if (l > -1 && ((unsigned int) l) < b->size) {
 			b->used = l + 1;
 
 			break;
